pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-limiter>=3.8.0",
    "flask-wtf>=1.2.1",
    "gunicorn>=23.0.0",
    "nltk>=3.9.1",
    "pillow>=10.4.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
    "pytesseract>=0.3.13",
    "redis>=5.1.1",
    "requests>=2.32.3",
    "tldextract>=5.3.0",
]
